Added development requirements

+++ accounts/models.py
@@ -1,6 +1,5 @@
 from django.contrib.auth.models import AbstractUser
-
 from django.db import models
 
 class CustomUser(AbstractUser):
-    age = models.PositiveBigIntegerField(null=True, blank=True)
+    age = models.PositiveIntegerField(null=True, blank=True)
